Door logging now only logs if status is different than last
The main reason for this is so that we can keep sendingupdates to the server in case one of the transmissions was dropped.  only the ones that change status will be logged

Also moved processing door message logic out of processMsg() This is just to clean up and segregate code.